refactor/#94 - add DataProviderContext to application
